addrgen_write/kernels: Consolidate unicast senders, multicast senders, readers, and receivers
