+++ tests/tt_metal/tt_fabric/fabric_data_movement/addrgen_write/kernels/kernel_common.hpp
@@ -1,0 +1,11 @@
+// SPDX-FileCopyrightText: © 2025 Tenstorrent Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include <cstdint>
+
+// Enum definitions for compile-time parameters
+enum class OperationType : uint32_t { BasicWrite = 0, Scatter = 1, FusedAtomicInc = 2 };
+
+enum class ApiVariant : uint32_t { Basic = 0, WithState = 1, SetState = 2 };
