+++ tests/tt_metal/tt_fabric/fabric_data_movement/addrgen_write/unicast_runner.cpp
@@ -14,6 +14,7 @@
 #include "tests/tt_metal/tt_metal/common/multi_device_fixture.hpp"
 #include "tests/tt_metal/tt_fabric/common/utils.hpp"
 #include "tests/tt_metal/tt_fabric/fabric_data_movement/addrgen_write/test_common.hpp"
+#include "tests/tt_metal/tt_fabric/fabric_data_movement/addrgen_write/kernels/kernel_common.hpp"
 #include "tt_metal/tt_fabric/benchmark/collectives/common/perf_helpers.hpp"
 #include "tt_metal/fabric/fabric_context.hpp"
 #include <tt-metalium/global_semaphore.hpp>
@@ -227,40 +228,31 @@ Notes:
          p.api_variant == AddrgenApiVariant::FusedAtomicIncWriteWithState ||
          p.api_variant == AddrgenApiVariant::FusedAtomicIncWriteSetState);
 
-    // Determine if this is a scatter variant
-    const bool is_scatter =
-        (p.api_variant == AddrgenApiVariant::ScatterWrite ||
-         p.api_variant == AddrgenApiVariant::ScatterWriteWithState ||
-         p.api_variant == AddrgenApiVariant::ScatterWriteSetState);
-
     const std::string KDIR = "tests/tt_metal/tt_fabric/fabric_data_movement/addrgen_write/kernels/";
 
-    // Helper to select writer kernel based on API variant
-    auto get_writer_kernel_path = [&KDIR](AddrgenApiVariant variant) -> std::string {
+    // Helper to map API variant to OPERATION_TYPE and API_VARIANT compile-time parameters
+    auto get_operation_and_api_variant = [](AddrgenApiVariant variant) -> std::pair<OperationType, ApiVariant> {
+        // Returns OperationType and ApiVariant enums
         switch (variant) {
-            case AddrgenApiVariant::UnicastWrite: return KDIR + "unicast_tx_writer_cb_to_dst_addrgen.cpp";
-            case AddrgenApiVariant::UnicastWriteWithState: return KDIR + "unicast_tx_writer_with_state_addrgen.cpp";
-            case AddrgenApiVariant::UnicastWriteSetState: return KDIR + "unicast_tx_writer_set_state_addrgen.cpp";
-            case AddrgenApiVariant::FusedAtomicIncWrite: return KDIR + "fused_atomic_inc_tx_writer_addrgen.cpp";
+            case AddrgenApiVariant::UnicastWrite: return {OperationType::BasicWrite, ApiVariant::Basic};
+            case AddrgenApiVariant::UnicastWriteWithState: return {OperationType::BasicWrite, ApiVariant::WithState};
+            case AddrgenApiVariant::UnicastWriteSetState: return {OperationType::BasicWrite, ApiVariant::SetState};
+            case AddrgenApiVariant::FusedAtomicIncWrite: return {OperationType::FusedAtomicInc, ApiVariant::Basic};
             case AddrgenApiVariant::FusedAtomicIncWriteWithState:
-                return KDIR + "fused_atomic_inc_tx_writer_with_state_addrgen.cpp";
+                return {OperationType::FusedAtomicInc, ApiVariant::WithState};
             case AddrgenApiVariant::FusedAtomicIncWriteSetState:
-                return KDIR + "fused_atomic_inc_tx_writer_set_state_addrgen.cpp";
-            case AddrgenApiVariant::ScatterWrite: return KDIR + "scatter_tx_writer_addrgen.cpp";
-            case AddrgenApiVariant::ScatterWriteWithState: return KDIR + "scatter_tx_writer_with_state_addrgen.cpp";
-            case AddrgenApiVariant::ScatterWriteSetState: return KDIR + "scatter_tx_writer_set_state_addrgen.cpp";
-            default: TT_FATAL(false, "Unknown API variant"); return "";
+                return {OperationType::FusedAtomicInc, ApiVariant::SetState};
+            case AddrgenApiVariant::ScatterWrite: return {OperationType::Scatter, ApiVariant::Basic};
+            case AddrgenApiVariant::ScatterWriteWithState: return {OperationType::Scatter, ApiVariant::WithState};
+            case AddrgenApiVariant::ScatterWriteSetState: return {OperationType::Scatter, ApiVariant::SetState};
+            default: TT_FATAL(false, "Unknown API variant"); return {OperationType::BasicWrite, ApiVariant::Basic};
         }
     };
 
-    std::string receiver_kernel_name;
-    if (is_fused_atomic_inc) {
-        receiver_kernel_name = "fused_atomic_inc_rx_addrgen.cpp";
-    } else if (is_scatter) {
-        receiver_kernel_name = "scatter_rx_addrgen.cpp";
-    } else {
-        receiver_kernel_name = "unicast_rx_addrgen.cpp";
-    }
+    auto [operation_type, api_variant] = get_operation_and_api_variant(p.api_variant);
+
+    // All receivers use the unified kernel now
+    const std::string receiver_kernel_name = "rx_addrgen.cpp";
 
     auto rx_wait_k = tt::tt_metal::CreateKernel(
         receiver_prog,
@@ -286,25 +278,17 @@ Notes:
                       .set_page_size(CB_ID, p.page_size);
     (void)tt::tt_metal::CreateCircularBuffer(sender_prog, p.sender_core, cb_cfg);
 
-    // Reader kernel (DRAM->CB)
+    // Reader kernel (DRAM->CB) - now uses unified kernel with OPERATION_TYPE compile-time arg
     std::vector<uint32_t> reader_cta;
     tt::tt_metal::TensorAccessorArgs(*src_buf).append_to(reader_cta);
-    reader_cta.push_back(1u /*SRC_IS_DRAM*/);
+    reader_cta.push_back(static_cast<uint32_t>(operation_type));  // OPERATION_TYPE
+    reader_cta.push_back(1u);              // SRC_IS_DRAM
     reader_cta.push_back(NUM_PAGES);
     reader_cta.push_back(p.page_size);
 
-    std::string reader_kernel_name;
-    if (is_fused_atomic_inc) {
-        reader_kernel_name = "fused_atomic_inc_tx_reader_to_cb_addrgen.cpp";
-    } else if (is_scatter) {
-        reader_kernel_name = "scatter_tx_reader_to_cb_addrgen.cpp";
-    } else {
-        reader_kernel_name = "unicast_tx_reader_to_cb_addrgen.cpp";
-    }
-
     auto reader_k = tt::tt_metal::CreateKernel(
         sender_prog,
-        KDIR + reader_kernel_name,
+        KDIR + "tx_reader_to_cb_addrgen.cpp",  // Unified reader kernel
         p.sender_core,
         tt::tt_metal::DataMovementConfig{
             .processor = tt::tt_metal::DataMovementProcessor::RISCV_0,
@@ -313,15 +297,17 @@ Notes:
             .defines = defines});
     tt::tt_metal::SetRuntimeArgs(sender_prog, reader_k, p.sender_core, {(uint32_t)src_buf->address()});
 
-    // Writer kernel (CB->Fabric->dst + final sem INC)
+    // Writer kernel (CB->Fabric->dst + final sem INC) - now uses unified kernel with compile-time args
     std::vector<uint32_t> writer_cta;
     tt::tt_metal::TensorAccessorArgs(*dst_buf).append_to(writer_cta);
-    writer_cta.push_back(NUM_PAGES);
-    writer_cta.push_back(p.page_size);
+    writer_cta.push_back(static_cast<uint32_t>(operation_type));  // OPERATION_TYPE
+    writer_cta.push_back(static_cast<uint32_t>(api_variant));     // API_VARIANT
+    writer_cta.push_back(NUM_PAGES);       // TOTAL_PAGES
+    writer_cta.push_back(p.page_size);     // PAGE_SIZE
 
     auto writer_k = tt::tt_metal::CreateKernel(
         sender_prog,
-        get_writer_kernel_path(p.api_variant),  // Use helper to select kernel
+        KDIR + "unicast_tx_writer_addrgen.cpp",  // Unified unicast writer kernel
         p.sender_core,
         tt::tt_metal::DataMovementConfig{
             .processor = tt::tt_metal::DataMovementProcessor::RISCV_1,
